requirements.txt
google-cloud-pubsub
firebase-admin
pdfminer.six
transformers
torch
nltk
keybert
textblob
dotenv